mooingateway/src/test/javascript/spec/app/entities/mooinbase/tax-item/tax-item.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { TaxItemService } from 'app/entities/mooinbase/tax-item/tax-item.service';
import { ITaxItem, TaxItem } from 'app/shared/model/mooinbase/tax-item.model';

describe('Service Tests', () => {
  describe('TaxItem Service', () => {
    let injector: TestBed;
    let service: TaxItemService;
    let httpMock: HttpTestingController;
    let elemDefault: ITaxItem;
    let expectedResult: ITaxItem | ITaxItem[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(TaxItemService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new TaxItem(0, 'AAAAAAA', false, false, 0, 0, 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a TaxItem', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new TaxItem()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a TaxItem', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            isValued: true,
            isPercentage: true,
            value: 1,
            companyID: 1,
            taxID: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of TaxItem', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            isValued: true,
            isPercentage: true,
            value: 1,
            companyID: 1,
            taxID: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a TaxItem', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
